Holiday files refactored
